+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Any BREAKING CHANGE between minor versions will be documented here in upper case
 - Updated dependencies: `std`, `@types/react`, `esbuild`, `postcss-nesting`, `vento`.
 - Script runner now uses `/usr/bin/env bash` instead of `/bin/bash` to improve portability [#466].
 - `nav` plugin with non-pretty-urls pages [#467].
+- `picture` plugin: use encodeURI to handle filenames with spaces [#469], [#470].
 
 ## [1.18.4] - 2023-08-02
 ### Added
@@ -2365,6 +2366,8 @@ The first version.
 [#463]: https://github.com/lumeland/lume/issues/463
 [#466]: https://github.com/lumeland/lume/issues/466
 [#467]: https://github.com/lumeland/lume/issues/467
+[#469]: https://github.com/lumeland/lume/issues/469
+[#470]: https://github.com/lumeland/lume/issues/470
 
 [1.18.5]: https://github.com/lumeland/lume/compare/v1.18.4...HEAD
 [1.18.4]: https://github.com/lumeland/lume/compare/v1.18.3...v1.18.4