+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project try to adheres to [Semantic Versioning](https://semver.org/),
 but not always is possible (due the use of unstable features from Deno).
 Any BREAKING CHANGE between minor versions will be documented here in upper case.
 
-## [1.18.5] - Unreleased
+## [1.18.5] - 2023-09-01
 ### Added
 - JSON loader to esbuild [#473].
 - `includes` option to Vento plugin.
@@ -15,7 +15,7 @@ Any BREAKING CHANGE between minor versions will be documented here in upper case
 - Make Vento components synchronous.
 
 ### Fixed
-- Updated dependencies: `std`, `@types/react`, `esbuild`, `postcss`, `postcss-nesting`, `autoprefixer`, `vento`, `eta`, `lightningcss`, `liquidjs`, `preact`, `rehype`, `sass`.
+- Updated dependencies: `std`, `@types/react`, `esbuild`, `postcss`, `postcss-nesting`, `autoprefixer`, `vento`, `eta`, `lightningcss`, `liquidjs`, `preact`, `rehype`, `sass`, `minify_html`.
 - Script runner now uses `/usr/bin/env bash` instead of `/bin/bash` to improve portability [#466].
 - `nav` plugin with non-pretty-urls pages [#467].
 - `picture` plugin: use encodeURI to handle filenames with spaces [#469], [#470].
@@ -2376,7 +2376,7 @@ The first version.
 [#470]: https://github.com/lumeland/lume/issues/470
 [#473]: https://github.com/lumeland/lume/issues/473
 
-[1.18.5]: https://github.com/lumeland/lume/compare/v1.18.4...HEAD
+[1.18.5]: https://github.com/lumeland/lume/compare/v1.18.4...v1.18.5
 [1.18.4]: https://github.com/lumeland/lume/compare/v1.18.3...v1.18.4
 [1.18.3]: https://github.com/lumeland/lume/compare/v1.18.2...v1.18.3
 [1.18.2]: https://github.com/lumeland/lume/compare/v1.18.1...v1.18.2